Update slidebars.js - sb-style-overlay init error
PROBLEM: When page load, the Slidebars with "sb-style-overlay" class take a right margin when "sb-right" and a left margin when "sb-left". After open one first time, the error dead and the slidebars are correted.

SOLUTION: Apply the margin-[right/left] in the "transform: translate([value])" value.

Thanks for the plugin! :D